# frontend/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crystalith",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/notebooks/{notebook_id}/sessions/{session_id}/messages": {
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "Create Message",
        "operationId": "create_message_v1_notebooks__notebook_id__sessions__session_id__messages_post",
        "parameters": [
          {
            "name": "notebook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Notebook Id"
            }
          },
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "List Messages",
        "operationId": "list_messages_v1_notebooks__notebook_id__sessions__session_id__messages_get",
        "parameters": [
          {
            "name": "notebook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Notebook Id"
            }
          },
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageRead"
                  },
                  "title": "Response List Messages V1 Notebooks  Notebook Id  Sessions  Session Id  Messages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/notebooks/{notebook_id}/analysis": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Analyze Notebook",
        "operationId": "analyze_notebook_v1_notebooks__notebook_id__analysis_get",
        "parameters": [
          {
            "name": "notebook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Notebook Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/models": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "List Models",
        "description": "List all available AI models.\n\nOptionally filter by role (chat, embed, edit, autocomplete) or capability.",
        "operationId": "list_models_v1_models_get",
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "chat",
                    "embed",
                    "edit",
                    "autocomplete"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          },
          {
            "name": "capability",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Capability"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/models/{model_id}": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Get Model",
        "description": "Get a specific model by ID.",
        "operationId": "get_model_v1_models__model_id__get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisResult": {
        "properties": {
          "topics": {
            "items": {
              "$ref": "#/components/schemas/Topic"
            },
            "type": "array",
            "title": "Topics"
          },
          "relations": {
            "items": {
              "$ref": "#/components/schemas/Relation"
            },
            "type": "array",
            "title": "Relations"
          },
          "contradictions": {
            "items": {
              "$ref": "#/components/schemas/Relation"
            },
            "type": "array",
            "title": "Contradictions"
          }
        },
        "type": "object",
        "required": [
          "topics",
          "relations",
          "contradictions"
        ],
        "title": "AnalysisResult"
      },
...
      "Citation": {
        "properties": {
          "source_id": {
            "type": "integer",
            "title": "Source Id"
          },
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "chunk_id": {
            "type": "integer",
            "title": "Chunk Id"
          },
          "chunk_index": {
            "type": "integer",
            "title": "Chunk Index"
          },
          "page_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Number"
          },
          "paragraph_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paragraph Index"
          },
          "snippet": {
            "type": "string",
            "title": "Snippet"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "source_id",
          "source_name",
          "chunk_id",
          "chunk_index",
          "snippet"
        ],
        "title": "Citation"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MessageCreate": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant",
              "system"
            ],
            "title": "Role"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content"
          },
          "citations": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Citation"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Citations"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "MessageCreate"
      },
      "MessageRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant",
              "system"
            ],
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "citations": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Citation"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Citations"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "role",
          "content",
          "citations",
          "created_at",
          "updated_at"
        ],
        "title": "MessageRead"
      },
      "ModelRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "provider": {
            "type": "string",
            "enum": [
              "openai",
              "ollama"
            ],
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles",
            "description": "Model roles: chat, embed, edit, etc."
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities",
            "description": "Special capabilities: tool_use, image_input, etc."
          }
        },
        "type": "object",
        "required": [
          "id",
          "provider",
          "model",
          "display_name",
          "description"
        ],
        "title": "ModelRead",
        "description": "Response model for a single AI model."
      },
      "ModelsListResponse": {
        "properties": {
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelRead"
            },
            "type": "array",
            "title": "Models"
          },
          "default_chat": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Chat"
          },
          "default_embedding": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Embedding"
          }
        },
        "type": "object",
        "required": [
          "models"
        ],
        "title": "ModelsListResponse",
        "description": "Response model for the models list endpoint."
      },
...
      "Relation": {
        "properties": {
          "source_chunk_id": {
            "type": "integer",
            "title": "Source Chunk Id"
          },
          "target_chunk_id": {
            "type": "integer",
            "title": "Target Chunk Id"
          },
          "relation_type": {
            "type": "string",
            "enum": [
              "similar",
              "references",
              "contradicts"
            ],
            "title": "Relation Type"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "source_chunk_id",
          "target_chunk_id",
          "relation_type",
          "score"
        ],
        "title": "Relation"
      },
...
      "Topic": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "chunk_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Chunk Ids"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "chunk_ids",
          "keywords"
        ],
        "title": "Topic"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}